03-Agent/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
asyncer==0.0.7
beautifulsoup4==4.12.3
bidict==0.23.1
cachetools==5.5.2
certifi==2024.8.30
chainlit==1.3.0rc1
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.4
filetype==1.2.0
frozendict==2.4.6
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
gspread==6.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
Lazify==0.4.0
literalai==0.0.623
lxml==5.3.0
marshmallow==3.23.1
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.57.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==23.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
plotly==5.24.1
proto-plus==1.26.1
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
syncer==2.0.3
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.28.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
webencodings==0.5.1
websockets==13.1
wrapt==1.17.0
wsproto==1.2.0
yfinance==0.2.44
zipp==3.21.0